refactor(dialog-effects): move BlobError to blob/error.rs; dialog_common::test
Move the BlobError enum and its From impls into a dedicated blob/error.rs
module (re-exported, so `dialog_effects::blob::BlobError` is unchanged), and
switch the blob capability-path tests from `#[test]` to `#[dialog_common::test]`
so they run on wasm too (the module already opts into a dedicated worker).
